Disable recursiveEval() until we have a less error-prone mechanism than passthroughSetters()
